Build 3.1.5662.24906
Build 3.1.5662.24906
*  Fixed an issue where copy and invention usage was being multiplied by
the number of runs incorrectly, resulting in large differences in IPH
and profit calculations with large number of runs.
*  Prices for amarr and gallente starship engineering datacores will now
show correctly on the datacores tab
*  Auto num bps updates when entering runs should no longer give large
numbers for T1 items
*  The BPC needed for invention has been added to a tool tip for
Invention chance for T2 items
*  Fixed an issue where the amounts for components and raw materials
(not build/buy) were the same in all cases. Now the correct differences
will be displayed.