Code updates and reworked Shopping List price updates
Shopping List stats should now be the same as the BP tab  (if one item)
but the only difference is that we will show how the broker fees for
setting up a buy order will change the iph and profit – that’s it.
Removed all the weird processing of old for trying to recalculate
everything. Note if the user updates the quantity of the items, then the
costs, profits, and IPH will be off a little bit (not much) but I'm not
going to get this exact unless we rebuild the entire BP and reset them.