Build 3.1.5671.15106
*  Updated to Aegis_1.0_114220 Database and Images
*  Industry jobs will now update for all checked alt accounts in the
list without needing to load the character.
*  All forms should now scale properly with different windows settings.
Note, this is still an ongoing issue and will update as necessary if
users are still having issues.
*  The total cost on the shopping list now reflects the total cost of
the shopping list and not the original cost entered. If users update the
items, the cost will relfect those changes.
*  Removed taxes and fees check boxes for items on the shopping list.
The shopping list is intended for shopping, not selling. Future updates
may incorporate taxes for selling items but for now they are removed.
*  Added invention and copy costs (does not include usage) to the
shopping list.
*  Blue highlights on records users can invent (they have the T1 BP) and
T3 Invented items should now show for both All Blueprints and Owned
Blueprints and show the correct settings selected.
*  Fixed an issue where invention and copy materials were not updated
correctly when changing values of items in the shopping list.
*  Items that have components built with compoents (e.g. Jump
freighters) now display the correct amount of materials to build all
levels of components on the BP list. The lists will now contain the
correct amount of buy materials, and the raw and component costs for all
levels and will be the same when using build/buy.
*  Made a small adjustment to build/buy calcuations to take into account
the total build cost of the component (usage, etc) instead of just the
material cost for making a build or buy determination.
*  User settings (Default ME/TE and Suggest Build BP if Not Owned) were
sometimes not set correctly in blueprint calculations; this has been
corrected.