Fixed copy output error and small shopping list fix
The copy output from the shopping list and bp tabs should contain a CRLF
now after the total price. Left it off on accident.

Also, made sure to check the Location for comparing shopping list items.

Finally, the total items in the list will now reset to 0 when the clear
list button is pressed.