Build 3.1.5732.29083
Updated DB and Images to Galatea_1.0_114618
System Indexes will now show next to the system name in the Facility
System Drop-down.
All Ore Anomalies listed in the Industry Upgrade Belts tool will now
correctly show the number of asteroids, types, and take into account the
effect of truesec.
The number of turrets available in the Mining tab for the Prospect has
been corrected to the number in game.
Fixed an issue where the program would not correctly load a facility
with POS settings on startup.
You can no longer save a POS facility on manufacturing screen without
selecting system.
The radio button for Copy/Paste types on the BP tab will now save
correctly when 'Save Settings' is pressed.
Added error processing to facility xml loads that caused the program to
crash on startup.
Added the number and max of Research Jobs to the number and max of
Production Jobs on the Industry Jobs tracker screen.
BPCs have been added as items in the Item Prices tab for updating prices
of BPCs used in invention. New selectors for BPCs have been added to the
Item Prices tab and Assets window.
Corrected several issues with in-line updating of items and components
on the Shopping List.
Changed the behavior of the default flag on facilities - now if it's not
the default, it will dim grey and if the default, it will highlight
blue. At any time, users can double-click the default flag to load the
currently saved default facility.
Saving a facility on the manufacturing tab should now correctly set the
default flag.
The SSV format should now correctly save with commas in the place of
decimals when using the Copy to Clipboard button on the BP Tab.