Build 3.1.5754.17095
* Updated Db and Images to Vanguard_1.0_114986
* Added a new setting that outputs ingame links for items listed in your
copy/paste text from the shopping list or BP tab. The default setting is
no links.
* Corrected an issue where the option to Ignore Market Price for Base
Materials in the Reactions tab wasn't being taken into account for
calculations.
* Corrected an issue where refining Unrefined Reactions was not taking
into account the refinery tax and efficiency.
* Added Refinery tax, efficency, and corp standings to the Reactions
tab, which will save independently with this tab.
* Removed refinery settings in the program settings. All forms that use
refinery settings will be saved with the settings on that form.
* Corrected an issue where Wormhole systems would not show in the
facility drop downs. Note, the cost index for the system is still not
available for Wormhole Systems.
* Fixed an issue where the type of shopping list export would not save
with settings correctly.
* Build volumes when copying and pasting a shopping list to clip board
should now be correct compared to what is shown on screen.
* Users can now manually update the system index on the BP tab for the
selected index and activity. Note, updating the indexes with CREST will
overwrite manual entries if there is a value from the CREST Update.
* Copy and Invention times should now correctly take into account skills
and the number of laboratory lines.