Fixed grid sorting error - patch 3.1.5796.22059
Issue happened when the sort class was set to the listview
ListViewItemSorter property, then the user sorted the column, then tried
to refresh the grid where .Items.Add was called and it tried to sort
before the subitems were added.

I reordered all list view populations to set the subitems first, then do
the final add of the listview item so that the sorting functions will
keep their original sort settings (from the user) when the data is
refreshed as well as not crash. Just in case, I added a try/catch to the
compare function for the listivewitemcomparer to catch any areas I might
have missed.

Note, all LP store stuff is work in progress and only viewable in
developer mode.