Fix for counting T3 hulls twice in material costs
The intention here looks to be only to include datacores, so I change it
to only pick datacores explicitly. There is some danger that other input
material types were making it into the list, but the correct way to fix
that would be to select for them specifically, like we do later on for
T3 relic hulls.

As it was, it was also adding T3 relic hulls in this selection, and then
doing it again later on in the T3 specific section.