Update 3.1.5832.14128
Bug fixes:
* The radio button on the BP tab now will correctly select the type of
copy based on the settings from the manufacturing tab when double
clicking an item there.
* Rigs will show as Rig and not Module in Manufacturing tab.
* Columns should no longer show NaN for some price fields in the
Manufacturing tab.
* Users should now be able to update the shopping list with assets from
more than 1000 locations without crashing the program.
* Tax rates for Outposts should now save and be used in calculations
correctly.
* When an industry job hits 0 seconds on the Industry Jobs window, the
row should now update to say 'Complete' and not being a negative count
down.
* When selecting a system for capital component building, the drop down
will no longer have four copies of the system name.
* Copy times should now be correct for using more than one lab line.
* SVR * IPH will now sort correctly in the Manufacturing tab. Note: the
value shown is to two decimals and this value may be smaller than that
but still non-zero.
* Saving 'All Items' on the Asset window will now save properly and
reload when used.

New Stuff
* Users can now update the price of an item on the BP tab by clicking on
the box below 'Market Price' label. When this is updated, it is saved
for the entire program until updated again.
* FW system bonuses can now be set on the BP tab in Options and selected
for items in the Manufacturing tab. On the BP tab, the values are saved
for the system and will apply when using that system for calculations.
On the manufacturing tab, the values are added to the appropriate
facility (Manufacturing, Copying, etc.) in calculations.
* All sorting options have been removed from the following tabs and are
now saved with settings based on the column clicked by the user:
Blueprints, Update Prices, Manufacturing, Datacores, and Mining. Others
will be updated as necessary. Note: all lists in the program are now
sortable by clicking on the column to sort.
* The Endurance and various Ice Miners for the Prospect and Endurance
have been added to the mining tab. Both the Prospect and Endurance are
available for selection for Ice Mining.
* Added Capital Component facility fields to the Manufacturing tab.
* Added FW level fields used for each facility to the Manufacturing tab.
* Industry jobs will update all selected API's on first load
automatically now.

General
* Made improvements to updating the industry facilities that should
result in faster update times.