Build 3.3.6043.38735
* Updated IPH and Updater to use .Net 4.6.1 - please update your .Net
installation to ensure IPH runs correctly here:
https://www.microsoft.com/en-us/download/details.aspx?id=49981
* Updated database and images to YC_118.6_1.1 - this has fixed issues
with incorrect materials for Cenobite II's and Citadels.
* The mining tab now applies an isk charge to refining instead of
minerals returned as changed with the Citadel patch. 6.67 standing is
still required to get refining with no tax applied.
* Adding only invention/copy materials to the shopping list from the BP
Tab will add the materials even if adding invention/copy materials to
shopping list is turned off.
* Updating the ME on the BP tab no longer crashes the program.
* Double-clicking on an item to add to shopping list from the
manufacturing tab no longer crashes the program.
* R.A.M.s will now show the correct build quantity for build/buy and
correct buy quantity when buying is cheaper on the BP tab. When adding
an item to the shopping list, the correct amounts will be inserted as
well.
* The correct BP Build Time and Total Build Time will now show when the
invention time or copy time checks are selected. This had an effect of
decreasing the IPH when more runs were added.
* The Manufacturing list will now properly refresh when no items found.